Added Room Random Generation
Created a random Generation System that works with all rooms.
The exception to this is the Kitchen Scene which is currently missing?
Shall fix ASAP.

Also restructured the folders massively to be more useful. PLEASE take
note of this guys, as we don't need folders for all scripts, unity has a
function which will show you them. Things are now ordered by what
objects they relate to; all enemy images & code is now in the Enemies
folder, all Powerup images & code is now in the Powerups folder etc.